move message handler to its own module
